feat: Update navbar component links and icons
This commit updates the links and icons in the navbar component. The "Projects" link is changed to "About" and the corresponding icon is updated to represent a user. Similarly, the "About" link is changed to "Projects" and the icon is updated to represent a tool. These changes were made to improve the navigation and user experience of the website.